If --prefix is specified, don't create gfarm.conf in $CONFIG_PREFIX/$sysconfdir, rather, create it just in $CONFIG_PREFIX/etc.
i.e. If sysconfdir=/home/soda/etc, and if --prefix is not specified,
gfarm.conf will be created in /home/soda/etc.  But if --prefix /var/tmp
is specified, now it will be created in /var/tmp/etc. It would be created
in /var/tmp/home/soda/etc before this change.