merge r8993 from the 2.5 branch:
fix a critical problem #754 - sharedsecret configuration allows unauthenticated access in a certain case

this problem was caused by the change for #740 - sharedsecret may not retry to connect to gfmd, when gfmd is restarted,
and found by tatebe@-san.
this happens when getpwnam_r(3) returns 0 with pwd == NULL,
and this cause is found by tatebe@-san, too.